Eureka!!!
Ian, I was able to finish the code that allows for a system with N
planets(determined by initial conditions)! I have some test code that is
proof of concept so if you wanna put it to the test each cell is a
different easily understandable plot!!! my favorite is the one of the
earth and Sun with actual real world values! I'm so tired... good night